look_ai_project/requirements.txt
fastapi
uvicorn
openvino
numpy
pillow
